requirements/base.txt
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
psycopg2-binary==2.9.9
django-environ==0.11.2
boto3==1.34.144
Pillow==10.4.0
gunicorn==22.0.0
